+++ docs/changes.rst
@@ -8,6 +8,7 @@ Not yet released.
 **Improvements**
 
 * Docker container accepts :envvar:`WEBLATE_REMOVE_ADDONS` and :envvar:`WEBLATE_ADD_MACHINERY` to customize automatic suggestion services.
+* Added OpenMetrics compatibility for :http:get:`/api/metrics/`.
 
 **Bug fixes**